studymap_backend/requirements.txt
django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
Pillow
PyMuPDF
python-docx
python-pptx
pytesseract
openai
requests